Added Grid and Footer
Created a component to create and render a grid.
The size of the grid cells change based on the number of cells and size
of screen, but are always square.
Added a random start location in the top half of the grid, and random
end location in bottom half.
Added random walls on approx 20% of cells.